web-app/requirements.txt
Flask==2.0.2
pymongo==3.12.0
pika==1.3.1
python-dotenv==0.19.2
Werkzeug==2.0.2
